+++ src/render/vulkan/shader/geo3d_zpre.vert
@@ -13,7 +13,7 @@ layout(location = 6)  in vec4   weights;
 layout(location = 7)   in mat4  model;
 layout(location = 11)  in mat4  model_inv;
 
-layout(location = 15) in uvec2 id;
+layout(location = 15) in vec2  id;
 layout(location = 16) in uint  material_idx;
 layout(location = 17) in uint  draw_edge;
 layout(location = 18) in uint  joint_count;
